requirements.txt
requests
mysql-connector-python
python-dotenv
pandas
apache-airflow
